requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
dj-database-url==2.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
razorpay==1.4.1
django-filter==23.3
djangorestframework-simplejwt==5.3.0
whitenoise==6.6.0
gunicorn==21.2.0
drf-yasg==1.21.7
celery==5.3.5
redis==5.0.1
django-redis==5.4.0
drf-nested-routers==0.93.4
scikit-learn==1.3.2
numpy==1.26.0
scipy==1.11.3
pandas==2.1.1